requirements.txt
streamlit==1.54.0
pandas>=2.2.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
groq>=0.18.0
langgraph==0.4.0
langchain-groq==1.1.1
Pillow>=10.0.0
pymupdf>=1.24.0
tiktoken==0.9.0
langchain
langchain-core